+++ test/templates/vuejs-default/src/main/webapp/app/router/account.ts
@@ -1,0 +1,43 @@
+import { Authority } from '@/shared/security/authority';
+
+const Register = () => import('../account/register/register.vue');
+const Activate = () => import('../account/activate/activate.vue');
+const ResetPasswordInit = () => import('../account/reset-password/init/reset-password-init.vue');
+const ResetPasswordFinish = () => import('../account/reset-password/finish/reset-password-finish.vue');
+const ChangePassword = () => import('../account/change-password/change-password.vue');
+const Settings = () => import('../account/settings/settings.vue');
+
+export default [
+  {
+    path: '/register',
+    name: 'Register',
+    component: Register,
+  },
+  {
+    path: '/account/activate',
+    name: 'Activate',
+    component: Activate,
+  },
+  {
+    path: '/account/reset/request',
+    name: 'ResetPasswordInit',
+    component: ResetPasswordInit,
+  },
+  {
+    path: '/account/reset/finish',
+    name: 'ResetPasswordFinish',
+    component: ResetPasswordFinish,
+  },
+  {
+    path: '/account/password',
+    name: 'ChangePassword',
+    component: ChangePassword,
+    meta: { authorities: [Authority.USER] },
+  },
+  {
+    path: '/account/settings',
+    name: 'Settings',
+    component: Settings,
+    meta: { authorities: [Authority.USER] },
+  },
+];
