+++ test/templates/vuejs-default/src/main/webapp/app/router/index.ts
@@ -3,28 +3,16 @@ import Component from 'vue-class-component';
 Component.registerHooks([
   'beforeRouteEnter',
   'beforeRouteLeave',
-  'beforeRouteUpdate' // for vue-router 2.2+
+  'beforeRouteUpdate', // for vue-router 2.2+
 ]);
 import Router from 'vue-router';
-const Home = () => import('../core/home/home.vue');
-const Error = () => import('../core/error/error.vue');
-const Register = () => import('../account/register/register.vue');
-const Activate = () => import('../account/activate/activate.vue');
-const ResetPasswordInit = () => import('../account/reset-password/init/reset-password-init.vue');
-const ResetPasswordFinish = () => import('../account/reset-password/finish/reset-password-finish.vue');
-const ChangePassword = () => import('../account/change-password/change-password.vue');
-const Settings = () => import('../account/settings/settings.vue');
-const JhiUserManagementComponent = () => import('../admin/user-management/user-management.vue');
-const JhiUserManagementViewComponent = () => import('../admin/user-management/user-management-view.vue');
-const JhiUserManagementEditComponent = () => import('../admin/user-management/user-management-edit.vue');
-const JhiConfigurationComponent = () => import('../admin/configuration/configuration.vue');
-const JhiDocsComponent = () => import('../admin/docs/docs.vue');
-const JhiHealthComponent = () => import('../admin/health/health.vue');
-const JhiLogsComponent = () => import('../admin/logs/logs.vue');
-const JhiAuditsComponent = () => import('../admin/audits/audits.vue');
-const JhiMetricsComponent = () => import('../admin/metrics/metrics.vue');
-/* tslint:disable */
-// jhipster-needle-add-entity-to-router-import - JHipster will import entities to the router here
+import { Authority } from '@/shared/security/authority';
+const Home = () => import('@/core/home/home.vue');
+const Error = () => import('@/core/error/error.vue');
+import account from '@/router/account.ts';
+import admin from '@/router/admin.ts';
+import entities from '@/router/entities.ts';
+import pages from '@/router/pages.ts';
 
 Vue.use(Router);
 
@@ -49,98 +37,9 @@ export default new Router({
       component: Error,
       meta: { error404: true }
     },
-    {
-      path: '/register',
-      name: 'Register',
-      component: Register
-    },
-    {
-      path: '/account/activate',
-      name: 'Activate',
-      component: Activate
-    },
-    {
-      path: '/account/reset/request',
-      name: 'ResetPasswordInit',
-      component: ResetPasswordInit
-    },
-    {
-      path: '/account/reset/finish',
-      name: 'ResetPasswordFinish',
-      component: ResetPasswordFinish
-    },
-    {
-      path: '/account/password',
-      name: 'ChangePassword',
-      component: ChangePassword,
-      meta: { authorities: ['ROLE_USER'] }
-    },
-    {
-      path: '/account/settings',
-      name: 'Settings',
-      component: Settings,
-      meta: { authorities: ['ROLE_USER'] }
-    },
-    {
-      path: '/admin/user-management',
-      name: 'JhiUser',
-      component: JhiUserManagementComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/user-management/new',
-      name: 'JhiUserCreate',
-      component: JhiUserManagementEditComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/user-management/:userId/edit',
-      name: 'JhiUserEdit',
-      component: JhiUserManagementEditComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/user-management/:userId/view',
-      name: 'JhiUserView',
-      component: JhiUserManagementViewComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/docs',
-      name: 'JhiDocsComponent',
-      component: JhiDocsComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/audits',
-      name: 'JhiAuditsComponent',
-      component: JhiAuditsComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/jhi-health',
-      name: 'JhiHealthComponent',
-      component: JhiHealthComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/logs',
-      name: 'JhiLogsComponent',
-      component: JhiLogsComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/jhi-metrics',
-      name: 'JhiMetricsComponent',
-      component: JhiMetricsComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    },
-    {
-      path: '/admin/jhi-configuration',
-      name: 'JhiConfigurationComponent',
-      component: JhiConfigurationComponent,
-      meta: { authorities: ['ROLE_ADMIN'] }
-    }
-    // jhipster-needle-add-entity-to-router - JHipster will add entities to the router here
+    ...account,
+    ...admin,
+    ...entities,
+    ...pages
   ]
 });
